melodies can play with angular, find way to connect songs uniquely to melodies.
