+++ modules/templates/test.hbs
@@ -1,6 +1,7 @@
 <div class="">
   <div>
-    {{$ctrl.title}}
-    <p ng-click="$ctrl.playTri()">PLAY</p>
+    <!-- recognizes only the controller in {{}} -->
+    {{$ctrl.result}}
+    <button ng-click="$ctrl.playTri()">Play</button>
   </div>
 </div>
