+++ views/profile.hbs
@@ -37,6 +37,8 @@
                 SONG 1
                 {{datasong.song_title}}
                 <prof></prof>
+                <button ng-click="$ctrl.playTri()"></button>
+                <p></p>
               </figure>
               <figure>
                 SONG 2
@@ -53,13 +55,6 @@
         </div>
       </div>
     </main>
-
-    <!-- <script src="//ajax.googleapis.com/ajax/libs/jquery/1.6.4/jquery.min.js"></script>
-    <script src="../routes/profile.js"></script>
-    <script type="text/javascript" src="../modules/components/app.module.js"></script> -->
-    <!-- <script src="../dependencies_two/jquery.min.js"></script>
-    <script src="../routes/profile.js"></script>
-    <script src="../dependencies_two/less.min.js"></script> -->
